task1-CRUD/crud_app/books/views.py
# from django.shortcuts import render
from rest_framework import generics
from .models import Book, Author
from .serilaizers import BookSerializer, AuthorSerializer

# Create your views here.
class BookList(generics.ListCreateAPIView):
    queryset = Book.objects.all()
    serializer_class = BookSerializer

class BookDetail(generics.RetrieveUpdateDestroyAPIView):
    queryset = Book.objects.all()
    serializer_class = BookSerializer

class AuthorList(generics.ListCreateAPIView):
    queryset = Author.objects.all()
    serializer_class = AuthorSerializer